% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_bars.R
\name{display_bars}
\alias{display_bars}
\title{Display individual color palettes in saguaRo}
\usage{
display_bars(name)
}
\arguments{
\item{name}{The name of color palette in saguaRo. Use `names(saguaro_pal)`
to list all valid names.}
}
\value{
An image of the requested color palette in saguaRo
}
\description{
Display individual color palettes in saguaRo
}
\examples{
display_bars("bloom")
display_bars("coyote")

}
