% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_saguaro.R
\name{display_saguaro}
\alias{display_saguaro}
\title{Display individual color palettes in saguaRo}
\usage{
display_saguaro(name)
}
\arguments{
\item{name}{The name of color palette in saguaRo. Use `names(saguaro_pal)`
to list all valid names.}
}
\description{
Display individual color palettes in saguaRo
}
\examples{
display_saguaro("gila")

}
