% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saguaro.R
\name{plot_saguaro}
\alias{plot_saguaro}
\title{Makes simple graphs to visual available color palettes in saugaRo.}
\usage{
plot_saguaro(name = "bloom", plot = "box", n = length(saguaro(name)))
}
\arguments{
\item{name}{The name of color palette in saugaRo. Use `names(saugaro_pal)`
to list all valid names.}

\item{plot}{The name of pre-made plot to visualize indicated color palette.
Valid options are: box, circle, density, gradient.}

\item{n}{An integer limiting the number of colors used from the selected
color palette.}
}
\description{
Makes simple graphs to visual available color palettes in saugaRo.
}
\examples{
plot_saguaro()
plot_saguaro("oriole", "gradient")
plot_saguaro("mariposa", "density")
plot_saguaro(plot = "circle")
plot_saguaro(n = 3)

}
