% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saguaro.R
\name{saguaro}
\alias{saguaro}
\title{Makes the color palettes from saguaRo available as R palettes.}
\usage{
saguaro(name, n)
}
\arguments{
\item{name}{The name of color palette in saugaRo. Use `names(saugaro_pal)`
to list all valid names.}

\item{n}{An integer limiting the number of colors used from the selected
color palette.}
}
\description{
Makes the color palettes from saguaRo available as R palettes.
}
\examples{
saguaro("bloom")
}
